refactor: tool shared
